[BE][4/n] split pipeline_llama into a separate file
[ghstack-poisoned]
